Core: Add some helper methods to VcsManager
Make VcsManager provide a generic way to list all IVersionControls
and to search for a specific one based on its id.

At some point we want to register the IVersionControls directly
with the VcsManager instead of going through the object pool and this
is a first step into this direction.

Change-Id: Id4ded85b29487f6d8adcd0cec1de490afae07132
Reviewed-by: Orgad Shaneh <orgads@gmail.com>
Reviewed-by: Tobias Hunger <tobias.hunger@theqtcompany.com>