Add IsoIconsVar to QmakeProjectManager::QmakeVariable.
It uses "ISO_ICONS" as its variable name.

This is necessary for the ISO Icon Browser Plugin. A previous attempt
at adding a more generic API to QmakeProFileNode that would allow for
reading and writing custom qmake variables was not sufficient. This is
because qtquickdesigner plugins do not get created until design mode
is entered, meaning that it was too late to register custom qmake
variables via static functions in QmakeProFileNode.

Change-Id: I2c8e0d944a1b9f07d1a7a70bd5e02ea371cf7680
Reviewed-by: Daniel Teske <daniel.teske@theqtcompany.com>