Fix for import management
If we add a new import we have to scan for new item library entries
in the directory belonging to the import.

Without this patch e.g. the commercial controls were not added
to the item library. The user had to switch to edit mode and back
to trigger and update.

Change-Id: Ie145aea53fbd57ffac976725b6c1fd5553ddf22e
Reviewed-by: BogDan Vatra <bogdan@kde.org>