Android: Fix debugging applications that link many modules
When debugging an application we will add a few arguments
to the command line for launching the application. We would do
the same for every library etc. required by the modules linked
to the application. At some point, the command line would become
too long and fail to execute.

Passing the libraries, libs_prefix, etc. on the command line
is not necessary, since the parameters are already included in
the AndroidManifest.xml, otherwise it would not be possible to
launch the application from the device itself.

Change-Id: I93a7f64a4ce32ebb2b25c54b0d17ae2b23706e24
Task-number: QTCREATORBUG-13691
Reviewed-by: Daniel Teske <daniel.teske@theqtcompany.com>