MimeDatabase: Prevent funny mixing of different mime type definitions
If multiple definitions for the same mime type are found, the mime XML
parser/provider would on the one hand replace the first MimeType instance
by the later one, but on the other hand keep all previously parsed glob
and magic matchers from the first definition.

The patch
- makes the XML parser ignore definitions of mime types that are already
defined
- parses custom types first, so custom types can override types from
freedesktop.org.xml

This fixes opening Cmake projects (CMakeLists.txt) in Qt Creator,
because freedesktop.org.xml defines these files to be x-cmake, but
we need them to be x-cmake-project.

Change-Id: Ia5598c2d0201f608aed6ae8c3268f0512d822f51
Reviewed-by: Daniel Teske <daniel.teske@theqtcompany.com>