Timeline: Allow horizontal resizing of detail windows
The details table can become quite large, so we elide the text.
However, as we don't know what kind of information we're hiding and
the user might want to see it it's quite practical to allow resizing.

Also, make sure that the detail window (and the range window) is on
top of everything else so that it can set the cursor shape.

Task-number: QTCREATORBUG-14054
Change-Id: I12850124c327c57c148b69960687b7402bdead90
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>