Revert "Do not add an "Analyzer" category to the "Issues" pane filter."
This reverts commit 324854a5cf522d32b425bc7de79d485f34e3877f.

It turns out that while there are no current users of this category,
it does make sense and will in fact be used soon.
In particular, the fact that some analyzers have their own issue views
does not matter, as these are for "regular" errors that the analyzers
find while doing their job, while the general Issues pane would be for
"irregular" errors that occur running an analyzer.

Change-Id: Ib595ff0e0b27a62634c41ffba220b312f5442f44
Reviewed-by: hjk <hjk@theqtcompany.com>