Disable re-run ability for analyzers.
Our analyzers were designed with the assumption that only one of them is
ever running (e.g. there is only one instance of the respective UI
element for each Analyzer). Run controls, on the other hand, assume that
an arbitrary number of them can run concurrently. With "re-run" enabled
for an analyzer run control, these concepts clash. A quick test shows
that some analyzers actually crash if you try to re-run them, but even
if this were not the case, it could not work in a sensible way.
Perhaps it would make sense to change the analyzer design so that they
too can run concurrently, but not for 3.4, obviously.

Change-Id: Ie8650eeef0261f2b697269900d5b465aad10aaf2
Reviewed-by: hjk <hjk@theqtcompany.com>
Reviewed-by: Daniel Teske <daniel.teske@theqtcompany.com>