Create a dummy run configuration to create our run control from.
The Clang Static Analyzer differs from other analyzers in that it does
not run a binary produced by the build process, but looks at source
files instead. It is therefore completely unrelated to any run
configurations that may or may not exist for the project. This has been
ignored so far, with these two main consequences:
    - When running the analyzer, the name of some random run
      configuration appears in the application output pane, which makes
      it look to the user as if the corresponding executable has been
      run, which it has not.
    - For projects without run configurations (e.g. libraries),
      analyzing does not work out of the box, which makes no sense
      conceptually.
So we now create our own run special run configuration (not visible in
the UI) and run it directly via runRunConfiguration() instead of using
the currently active run configuration via runProject().
This fixes both issues listed above.

Change-Id: Icc839816f4a1e6f02a0eb2328c536b44f7304807
Reviewed-by: Nikolai Kosjar <nikolai.kosjar@theqtcompany.com>