Timeline: Use QHash instead of QVector for child render states
The offsets can get very large if you have long traces or zoom in very
closely. We hardly ever need more than a few on each level, so this is
just a waste of memory. Also, use 64bit integers to index them, as we
can get very close to the numerical limit with 32bit ones.

Change-Id: I20db4ea5dd2ea8922fa6552c106194bb4f19a76b
Task-number: QTCREATORBUG-14105
Reviewed-by: Kai Koehne <kai.koehne@theqtcompany.com>