Utils: Fix parentheses parsing in arguments iteration
Opening parenthesis is stored in the value, but closing parenthesis
is not. This makes the value unbalanced.

Since the parenthesis unconditionally unset m_simple, the value is
cleared anyway and is never read.

In case that math (or other) expression will ever be evaluated, this
needs to be reworked to include both parentheses.

Change-Id: Ie56f5a5d9cec5a47649751db4245888bee8e85d8
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>