MimeDatabase: Fix that xml files were opened in bar descriptor editor
Fix MIME detection issues with magics in MIME hierarchies

Assume two MIME types A and B are registered, both with the same glob
pattern, A being parent of B, A with some magic rule, and B with another
magic rule. Given a file that matches the glob pattern and the magic rule
of A, the resulting MIME type depended on the order of registration of A
and B, because it would just check if some glob matching MIME type was
also a subclass of the magic matching MIME type.

The patch prefers the the MIME type that matches by magic if that
matches by glob pattern as well (i.e. A in our example).

The "recommended checking order" of the spec does handle that case.

Change-Id: Ia914aa7b6d0fb52f6c833897a5be69eb59fca6ab
Task-number: QTBUG-44846
Reviewed-by: Daniel Teske <daniel.teske@theqtcompany.com>