MimeDatabase: Allow regular expressions for magic rules
This extends the freedesktop standard, which really is lacking
in many ways.
Adds a type "regexp", which interprets the "value" as a
QRegularExpression with MultilineOption and DotMatchesEverythingOption.
The range/offset is the range of the sub-string that is matched
against the regular expression. If the value for the end of the
range is omitted, or the same as the start, the whole string is
matched against the regular expression.
The mask is ignored.

Change-Id: Iaaaf769451ed23112f1fbeee254fdff3198b74a2
Reviewed-by: BogDan Vatra <bogdan@kde.org>