Vcs: Introduce VcsBaseClientImpl
Introduce VcsBaseClientImpl class that is a base for both GitClient
and VcsBaseClient.

The idea is to slowly move code from VcsBaseClient into VcsBaseClientImpl
and reuse that in GitClient. Ideally GitClient would become a full
VcsBaseClient at some point, but let's see how far we can get with this
conversion process:-)

Change-Id: I2e63ba0dcf8a61128d5abf7c1453625e1b9f4e7e
Reviewed-by: Orgad Shaneh <orgads@gmail.com>