Timeline: Remove useless dataAvailable() signals
The only thing they did was trigger some QML code that set the zoom to
10% of the available range. 10% is somewhat arbitrary, so we shouldn't
have it in the general purpose timeline. Also, we don't really have to
pass the signal through several connections just to do that in QML.

The clearChildren() didn't have any effect there as dataAvailable()
should only happen if the view is already cleared.

Change-Id: I1813ed37a82fc3fd5c3b7fc02b2afb1dcbcadcfb
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>