Fix parsing of encoded xml attributes
If an xml attribute contains an entity parsing it failed due to
missing decoding. This patch decodes entities holding (hexa-)decimal
entities.
This is especially necessary for files (or paths) containing some
special characters that might end up encoded inside the output that
would be generated by running the tests.

Change-Id: I4f3b9f9edc59ff1433b92ed4ce1933eaf29ffe74
Reviewed-by: David Schulz <david.schulz@theqtcompany.com>