Timeline: Properly encapsulate render pass states
We have to make sure that all the nodes get deleted when the states are
deleted. Previously, we relied on the RenderState destructor to
recursively delete nodes owned by their parents. This is rather hard to
understand and can easily fail if we create a pass state without calling
TimelineRenderState::assembleNodeTree() afterwards.

The best way to deal with this is to properly encapsulate the nodes into
the states and add destructors which delete them.

Change-Id: I8b1ce16084afc1c85a90609f8f8d889147f7832f
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>