QmlProfiler: Clear timeline models before clearing the base model.
The model manager's stateChanged() signal gives us accurate information
on when the clearing begins so that we can clear the timeline models as
first thing, not as last.

Clearing them as last thing can indirectly prompt them to access now
invalid data of the base model, such as the types.

Change-Id: I3af4a6716ed8b2efb433e7a264d3fc9a65f61490
Task-number: QTCREATORBUG-14277
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>