Clang: Fix providing clang directory with intrinsics
Opening a main.cpp of a Qt Widget Application project that is configured
with a GCC toolchain leads to errors [1] like

    /usr/lib/gcc/x86_64-linux-gnu/4.8/include/ia32intrin.h:41:10: \
    error: use of undeclared identifier '__builtin_ia32_bsrsi'

due to a wrong determination of the clang directory with intrinsics. The
directory was determined as

    $${LLVM_LIBDIR}/clang/$${LLVM_VERSION}/include

whereas $LLVM_VERSION was extracted as e.g. "3.5" from 'llvm-config
--version' instead of "3.5.0".

The path of clang intrinsics dir shipped with Qt Creator (as package)
does also match this version scheme:

    $QTC_INSTALL_DIR/share/qtcreator/cplusplus/clang/3.4.2/include

[1] Visible with qtc.clangcodemodel.verboserun=true

Change-Id: I1061912eef437a9889987e638915d2fabe208011
Reviewed-by: hjk <hjk@theqtcompany.com>
Reviewed-by: Eike Ziller <eike.ziller@theqtcompany.com>