KitManager: Enable sorting on the target setup page
Add a separate method to sort lists of kits and remove the sortedKits()
method again.

You should not use the sorted list whenever possible, so requireing a
two-step process to get that is not too hard. And it allows to sort
sub-sets of all kits, which is exactly what we might have on the
TargetSetupPage.

Task-number: QTCREATORBUG-13549
Change-Id: Iaa8443bcc8c321425daa13a1fb1e024d18939306
Reviewed-by: Daniel Teske <daniel.teske@theqtcompany.com>