Replace ListView by TreeView inside results pane
This removes the flickering of the view when adding too many items
in a short time frame. Additionally it is the first step towards
making the results a real tree for having a better overview and to
be able to reduce the output easier to what is wanted by the user.

Change-Id: Id515bf0f43c5037d956bdbf2818a5d6ef97c82dd
Reviewed-by: David Schulz <david.schulz@theqtcompany.com>