ProjectTree: Improve save/load of expand data
So far the code only saved the path of the expanded nodes, which always
created problems for nodes that had the same path. This broke for the
virtual folders SOURCES, FORMS, HEADERS recently because they have
nowadys the same path. They used to have different paths.

This new code saves both the path and the displayname and compares that,
this fixes the expanding for those folders.

Obviously there is no sure way to indetify the right nodes to expand,
and this isn't backwards compatible.

Change-Id: I9854f90ca942f92420970765b0cc19138ad5d63d
Task-number: QTCREATORBUG-14304
Reviewed-by: Tobias Hunger <tobias.hunger@theqtcompany.com>