Android: Increase timeout for checkPasswd and checkKeystore
The timeout was "only" 4s, but apparently pretty much every
android tool randomly takes longer than 4s.

Increasing this timeout won't affect anyone for whom the
android tool returns in 1s, but makes it work for anyone
where the checking takes longer.

This waiting though is done in the main thread, which is pretty
bad. Moving it to a background thread isn't easy though.

Change-Id: I135b1d1ee9634005d5eb6038a46d66b6a8e97fca
Task-number: QTCREATORBUG-13175
Reviewed-by: BogDan Vatra <bogdan@kde.org>