ProjectTree::currentProject: Introduce a fall back to startupProject
If no node is selected, e.g. no document is open, the ProjectTree
currently claims that there is no current project.

This adds a fallback to startup project in that case. This especially
fixes the case where only one project is open and as such it is clear
from context which project the user expects to be current.

Revert a few places where I implemented this fallback manually.

I do not like this. If multiple projects are open, this fallback can be
surprising and not what the user wants. Everything works fine in the
simple case with one project open but breaks with multiple projects
open, this actively hides bugs.

Change-Id: I6259834ca2220dd83bdffb16c3a0eac8f98e504b
Task-number: QTCREATORBUG-14394
Reviewed-by: Eike Ziller <eike.ziller@theqtcompany.com>