ProjectTree: Improve resizing of project tree
On each resize event the treeview iterated over all rows to get the
maximum width of them. That involves text layout and is very slow
with bigger trees.

So cache the size hint for the column and connect to all signals
that can potentially change the size hint.

Change-Id: I83a13982848df56515fed8225ced336096370546
Task-number: QTCREATORBUG-8075
Reviewed-by: David Schulz <david.schulz@theqtcompany.com>