ManhattanStyle: Fix size of field in form layouts on OS X
The default in QMacStyle, to use FieldsStayAtSizeHint, just almost
always is the wrong thing to do. Additionally this goes undetected for a
while most of the time, because most people do not run their code on OS
X. Override it globally to AllNonFixedFieldsGrow like in other shipped
styles.

Task-number: QTCREATORBUG-14337
Change-Id: I3b0719c0201871fbf0163b54b1a54823bf5d14f3
Reviewed-by: Daniel Teske <daniel.teske@theqtcompany.com>