FileSearch: Avoid copy of function object.
Converting std::unary_function to std::function will create a copy,
which we do not want. Actually std::unary_function as a base class
doesn't serve a purpose at all in this context, so we can just remove
it. Bind a std::function to a pointer to our search object instead.

Change-Id: I31db62312eefdff19b09fbb4eb20e8e666125874
Reviewed-by: Daniel Teske <daniel.teske@theqtcompany.com>