Fix crash at close after editor failed to open
When a file is part of a project, but fails to open in an editor, the
editor configuration would still keep a reference to the "created" editor
that in the end was not "opened".
Remove editors from the configured editors list when they are destroyed,
instead of when they are "closed".

Change-Id: Ib9680d56b737214c09f935f0088a77959ca7e8ba
Reviewed-by: Daniel Teske <daniel.teske@theqtcompany.com>