Squish: Fix verification in tst_rename_file
This doesn't change anything until we replace the used project.
main.cpp or utility.cpp may get the tempFiletype "Headers" and
then the test will crash because the file can't be opened.
In addition, only files with filetype "Headers" are included in
other files, so verifying including of other files is
pointless.

Change-Id: I30b20269a800eca5ef51de3af209e674da5189bd
Reviewed-by: Christian Stenger <christian.stenger@theqtcompany.com>