Resource editor: Remove lots of indirections
Because only the tree view had access to the resource model, many
methods in the IEditor instance and document were going their way
through IEditor -> QRC editor widget -> tree view -> model.
Create the model in the document instead, pass that model to the tree
view, and let the document work on the model directly.

Change-Id: I76405e60f118e2bbf63d3f9a4d39cd73be64aa14
Reviewed-by: Daniel Teske <daniel.teske@theqtcompany.com>