QtVersion: Add features for major/minor/patchlevel of Qt version
This is more fine-grained than the "Qt4" or "Qt5" we used to support
and will allow for wizards to specify things like "Make sure we get
a Qt version > 5.x" for the kit selection page.

That page is provided by the ProjectExplorer, so we can not access
the Qt version there, so Features seem the only possible way to go.

Change-Id: I0f4b3a0b5be91fb331924a560414e8f73bccca41
Reviewed-by: Orgad Shaneh <orgads@gmail.com>
Reviewed-by: Daniel Teske <daniel.teske@theqtcompany.com>