QmlProfiler: Always try to connect immediately if we know the port
Waiting for the debug output is not necessary in most cases because
we tell the application which port to listen on anyway.
QmlProfilerClientManager will retry 50 times and then prompt the user
to decide if even more tries are necessary. This should be enough to
establish a connection.

As a side effect, this removes the only place where
AnalyzerStartParameters::startMode is used in the QML profiler.

Change-Id: I40991b34c98d93b0874d627fba3dae7ccfabaf2c
Task-number: QTCREATORBUG-13510
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>
Reviewed-by: David Schulz <david.schulz@theqtcompany.com>