JsonWizard: Explicitly set required/preferred features on kits page
Do not use the required/preferred features from the wizard in the
kits page. The wizard's features are used to decide which wizards
will be shown (together with the enabled key). The kits page has
a separate set of features that are used to decide on the Kits
shown.

Keep these settings separate, as they are different things.

If nothing is set at the Kits page, then the wizard's values
will be copied. This is to stay compatible with existing
JSON wizards.

Change-Id: I2f8e3ccc46ce92f5be89aecaf52e5a81e4b29e83
Reviewed-by: Orgad Shaneh <orgads@gmail.com>