Timeline: Fix display of time labels for empty traces
If the trace being shown is shorter than 1ns we better don't show any
labels as those are necessarily useless. Trying to "repair" the
duration to show something in that case creates more problems than
it solves, as the re-evaluation of the bindings when the actual values
arrive produces strange transitional states, which surface as large
negative numbers for the time labels.

Change-Id: I16598a2b1c3271dc50bbd3877baf4c0e15cc6411
Reviewed-by: Christian Stenger <christian.stenger@theqtcompany.com>